adding test and fixing docs
