+++ src/qibo/tests/regressions/vqc_bfgs.out
@@ -1,0 +1,54 @@
+3.716941486648598580e+00
+4.523823071786376993e+00
+3.942900859047370421e+00
+3.476271704158994336e+00
+2.369968896767256084e+00
+3.864468408728399140e+00
+2.769911332093198730e+00
+5.598786224337186646e+00
+5.917284777841676835e+00
+2.392882483318851783e+00
+5.254908878101084646e+00
+3.435336570056210181e+00
+3.596790078124608936e+00
+5.827587143161817806e+00
+3.805373580431322122e-01
+5.885011661932784133e-01
+2.706234319376334407e-01
+5.367610000511117363e+00
+5.097041224326826381e+00
+5.389295296655916445e+00
+5.923705007203717621e+00
+4.938654860088706577e+00
+2.874489206323588508e+00
+4.766260782074382973e+00
+9.792116173329254991e-01
+4.144255778077049257e+00
+7.335124744852268286e-01
+5.904957483673502594e+00
+3.283089821693081944e+00
+2.423006216556704118e+00
+1.448031875765958487e+00
+4.809080208357267239e+00
+3.030095772947239308e+00
+3.455735990913264999e+00
+1.037823171274397938e-01
+3.824214885049912915e+00
+3.633800350169739790e+00
+3.848660123223391683e+00
+6.031574650094103873e+00
+4.280178742794634061e+00
+2.269733156716351452e+00
+3.044600442488948122e+00
+4.284135946645464266e+00
+5.089684007847219416e-01
+4.223577777684530687e+00
+4.143166555013832131e+00
+1.252918791038462043e+00
+9.624343040481694311e-01
+1.874535819937320325e+00
+2.350435058135704303e+00
+3.780268096010602541e+00
+2.602255080599798376e+00
+6.157148019035041386e+00
+6.175770661286785757e-01
