+++ src/qibo/tests/regressions/vqc_powell.out
@@ -1,0 +1,54 @@
+5.817858516044246642e+00
+5.070846911026931636e+00
+3.378838047499987063e+00
+4.283083389917655026e+00
+1.279881690182928633e+00
+4.682486741224203719e+00
+2.950418381991937178e+00
+5.213872900426652635e+00
+5.773794483046278181e+00
+2.381242725548876393e+00
+4.695155369292481673e+00
+3.303251398137756922e+00
+3.110410899084510739e+00
+5.836473773113795005e+00
+2.045648111007575265e-02
+1.529271742752940089e-01
+-7.190129591166213863e-02
+4.911932734720512173e+00
+4.190409229943121261e+00
+5.129075977969671740e+00
+6.172175008817991859e+00
+4.888693734143172875e+00
+2.825959502066920592e+00
+4.920510735096917720e+00
+7.882459018949533602e-01
+4.256485402021993991e+00
+5.286020423991792949e-01
+6.076515755375639039e+00
+3.591071636467368489e+00
+2.440665361905427488e+00
+1.553753420351141834e+00
+5.073692701607622979e+00
+3.827461624853559119e+00
+3.447158920251087721e+00
+-1.778623188492944873e-02
+3.281447757180323777e+00
+3.720675443681876970e+00
+3.920235676584029605e+00
+6.002851517129316861e+00
+4.232333471045549800e+00
+2.120265140162471251e+00
+2.757256233469632445e+00
+4.287193459345991720e+00
+2.630821352014618419e-01
+4.013388788543736929e+00
+4.546197684714632814e+00
+1.360224855591324111e+00
+8.199862620245802480e-01
+1.788695528116919897e+00
+2.400446090153833367e+00
+3.945995194154109154e+00
+2.789787578246364141e+00
+6.106755429444294947e+00
+8.367919690957653645e-01
